# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger documentation for Party API",
        "title": "Party API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/party": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "List parties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "party"
                ],
                "summary": "Get all upcoming parties",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Creates new party",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "party"
                ],
                "summary": "Create a new party",
                "parameters": [
                    {
                        "description": "Party Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                }
            }
        },
        "/party/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get a single party by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "party"
                ],
                "summary": "Get one party",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "party Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "deletes party",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "party"
                ],
                "summary": "Delete existing party",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "party Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "updates existing party",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "party"
                ],
                "summary": "Update existing party",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "party Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Party Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/party.Party"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "party.Party": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Maydan nezalezhnosti 1"
                },
                "city": {
                    "type": "string",
                    "example": "Kyiv"
                },
                "dateTime": {
                    "type": "string",
                    "example": "2022-12-31T23:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "New year party"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}